+++ CHANGELOG.md
@@ -268,6 +268,7 @@
 - Submitted reports can be moved into the review state 
 - Reports in review can be moved into the awaiting changes state   
 - Transactions & Planned Disbursements cannot be edited if they are associated with an approved Report
+- BEIS users can move a Report into the approved state
 
 - `Call open date` and `Call close date` added to the create activity form, for levels C and D.
   This field is mandatory for new activities, but optional for activities marked as `ingested: true`